// web-application/src/types-openapi/api/models/ScheduleRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface ScheduleRequest
 */
export interface ScheduleRequest {
  /**
   *
   * @type {number}
   * @memberof ScheduleRequest
   */
  staffId?: number | null;
  /**
   *
   * @type {number}
   * @memberof ScheduleRequest
   */
  shiftId?: number;
  /**
   *
   * @type {Date}
   * @memberof ScheduleRequest
   */
  startDate?: string | null;
  /**
   *
   * @type {Date}
   * @memberof ScheduleRequest
   */
  endDate?: string | null;
  /**
   *
   * @type {Array<string>}
   * @memberof ScheduleRequest
   */
  byDay?: Array<string> | null;
  /**
   *
   * @type {boolean}
   * @memberof ScheduleRequest
   */
  isFixedShift?: boolean;
}

/**
 * Check if a given object implements the ScheduleRequest interface.
 */
export function instanceOfScheduleRequest(value: object): value is ScheduleRequest {
  return true;
}

export function ScheduleRequestFromJSON(json: any): ScheduleRequest {
  return ScheduleRequestFromJSONTyped(json, false);
}

export function ScheduleRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ScheduleRequest {
  if (json == null) {
    return json;
  }
  return {
    staffId: json["staffId"] == null ? undefined : json["staffId"],
    shiftId: json["shiftId"] == null ? undefined : json["shiftId"],
    startDate: json["startDate"] == null ? undefined : json["startDate"],
    endDate: json["endDate"] == null ? undefined : json["endDate"],
    byDay: json["byDay"] == null ? undefined : json["byDay"],
    isFixedShift: json["isFixedShift"] == null ? undefined : json["isFixedShift"],
  };
}

export function ScheduleRequestToJSON(json: any): ScheduleRequest {
  return ScheduleRequestToJSONTyped(json, false);
}

export function ScheduleRequestToJSONTyped(value?: ScheduleRequest | null, ignoreDiscriminator: boolean = false): any {
  if (value == null) {
    return value;
  }

  return {
    staffId: value["staffId"],
    shiftId: value["shiftId"],
    startDate: value["startDate"] == null ? undefined : value["startDate"],
    endDate: value["endDate"] === null ? null : value["endDate"],
    byDay: value["byDay"],
    isFixedShift: value["isFixedShift"],
  };
}